Move User to a separate file, move MainState to its slice
