--- web/src/features/main/redux/state.ts
@@ -1,14 +1,0 @@
-import {Repository} from "../../repositories/Repository";
-import {User} from "../../login/AuthRepository";
-
-export interface MainState {
-    user?: User,
-    repositories: Repository[],
-    isLoading: boolean,
-    error?: string,
-
-}
-export const initialState: MainState = {
-    repositories: [],
-    isLoading: true,
-}
