+++ web/src/features/user/User.ts
@@ -1,0 +1,4 @@
+// TODO add more data mapping from backend and on backend too
+export interface User {
+    username: string
+}
